requirements.txt
ldap3==2.9.1
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1